requirements.txt
asgiref==3.2.7
bson==0.5.8
dataclasses==0.7
Django==2.2.12
django-cors-headers==3.2.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djongo==1.3.2
mongoengine==0.20.0
pkg-resources==0.0.0
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.2.4
